labeltoolv2.0/requirements.txt
absl-py==1.2.0
args==0.1.0
astor==0.8.1
certifi
click==7.1.2
clint==0.5.1
coco==0.4.0
docopt==0.6.2
ffmpeg==1.4
gast==0.5.3
grpcio==1.47.0
h5py==3.7.0
imageio==2.20.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mock==4.0.3
mrcnn==0.2
networkx==2.6.3
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.19.0
pyparsing==3.0.9
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyWavelets==1.3.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
scikit-image==0.19.3
scipy==1.7.3
shiboken6==6.3.1
six==1.16.0
tablib==3.2.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tifffile==2021.11.2
Werkzeug==2.2.0
wincertstore==0.2
zipp==3.8.1
